+++ examples/js/index.html
@@ -17,8 +17,8 @@
   <pre id="output">Click the button to run the demo...</pre>
 
   <!-- Include the bundled SDK -->
-  <!-- Adjust path if your example-js/index.html is served from a different relative location to sdk-js/dist -->
-  <script src="../sdk-js/dist/bundles/slipkey-sdk.umd.js"></script>
+  <!-- Adjust path if your example-js/index.html is served from a different relative location to slipkey-js/dist -->
+  <script src="../slipkey-js/dist/bundles/slipkey-sdk.umd.js"></script>
 
   <script>
     const outputEl = document.getElementById('output');
