+++ examples/js/package.json
@@ -8,7 +8,7 @@
     "start": "ts-node --esm src/example.ts"
   },
   "dependencies": {
-    "slipkey-sdk": "file:../sdk-js"
+    "slipkey-sdk": "file:../../sdk-js"
   },
   "keywords": [],
   "author": "",
