+++ examples/js/src/example.ts
