+++ examples/js/tsconfig.json
